feat: update basic_cleaning component with working code
Here I update the MLflow component for performing the basic cleaning of
the data. I took the initial files generated by the Cookiecutter and
filled in the required code and tested that it ran with mlflow.

The main.py file was also updated to run this step when the pipeline is
run. I also updated the cnfig.yaml file to include the additional
max_minimum_nights argument.